wview and wdelete : Add checking rules for index

+++ src/main/java/seedu/commands/workoutcommands/DeleteWorkoutCommand.java
@@ -33,9 +33,10 @@ public class DeleteWorkoutCommand extends Command {
     //@@author calebcjl
     @Override
     public String execute() throws InvalidArgumentException {
-        if (workoutToDeleteIndex >= workoutList.getWorkouts().size()) {
-            throw new InvalidArgumentException("index");
+        if (workoutToDeleteIndex >= workoutList.getWorkouts().size()|| workoutToDeleteIndex<1) {
+            throw new InvalidArgumentException("Index cannot be 0 or more than " + workoutList.getWorkouts().size());
         }
+
         String workoutName = workoutList.getWorkout(workoutToDeleteIndex).getWorkoutName();
         Date workoutDate = workoutList.getWorkout(workoutToDeleteIndex).getDate();
         workoutList.deleteWorkout(workoutToDeleteIndex);
